backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
google-genai==1.66.0
openai==1.30.0
httpx==0.28.1
python-dotenv==1.0.1
groq==0.9.0
pyyaml==6.0.1
